refactor: split oversized source files for readability
- translations.ts (845 lines) を types/ja/en に分割し、translations.ts は
  互換バレル（再エクスポート）に変更。import パスは従来どおり。
- nosskey.ts (906 lines) から未消費 PRF キャッシュ機構を PendingPrfCache
  クラス (pending-prf-cache.ts) へ抽出。TTL ゼロ化を含む自己完結したサブ
  システムを分離し、NosskeyManager の責務を明確化。PENDING_PRF_TTL_MS は
  既存 import 互換のため nosskey.ts から再エクスポート。
