+++ examples/svelte-app/src/i18n/translations/en.ts
@@ -1,0 +1,285 @@
+import type { TranslationData } from './types.js';
+
+// 英語翻訳
+export const en: TranslationData = {
+  common: {
+    save: 'Save',
+    cancel: 'Cancel',
+    close: 'Close',
+    ok: 'OK',
+    error: 'Error',
+    success: 'Success',
+    back: 'Back',
+    copy: 'Copy',
+    copied: 'Copied',
+    loading: 'Loading...',
+    edit: 'Edit',
+    help: 'Help',
+    errorMessages: {
+      init: 'Initialization Error:',
+      prfCheck: 'PRF Support Check Error:',
+      passkeyCreation: 'Passkey Creation Error:',
+      login: 'Login Error:',
+      importNsec: 'nsec Import Error:',
+      prfUnsupported:
+        "This passkey can't be used. Some authenticators such as Bitwarden are not supported, so please try a different passkey that supports the PRF extension (such as a platform passkey).",
+    },
+  },
+  appWarning: {
+    title: 'Important Notice',
+    prfCompatibility: 'Some devices and environments are not supported.',
+  },
+  auth: {
+    title: 'Nosskey',
+    subtitle: 'Your Nostr account, with a passkey',
+    createNew: 'Create',
+    loginWith: 'Login with Passkey',
+    loading: 'Loading...',
+    username: 'Username (Optional)',
+    usernamePlaceholder: 'Enter username',
+    tabLogin: 'Login',
+    tabRegister: 'Sign Up',
+    usernameTip: 'Used as the display name for your passkey. Optional — you can leave it blank.',
+    methodImport: 'Import secret key (Beta)',
+    nsecLabel: 'Existing nsec',
+    nsecPlaceholder: 'nsec1...',
+    nsecTip:
+      'Enter your existing Nostr private key (nsec1...). It will be encrypted with your passkey and stored locally; signing and encryption operations work transparently via the passkey.',
+    importNsec: 'Import',
+    invalidNsec: 'Invalid nsec format. Please enter a bech32 string starting with "nsec1...".',
+    accounts: {
+      title: 'Saved accounts',
+      relogin: 'Re-login to this account',
+      delete: 'Remove',
+      confirmDelete: 'Yes, delete',
+      cancel: 'Cancel',
+    },
+  },
+  nostr: {
+    publicKey: 'Public Key',
+    copyToClipboard: 'Copy to Clipboard',
+    copiedToClipboard: 'Copied',
+    profileAvatarAlt: 'Profile picture',
+    profileFallbackAlt: 'No avatar',
+  },
+  key: {
+    title: 'Key Management',
+  },
+  settings: {
+    title: 'Settings',
+    exportSecretKey: 'Export Secret Key',
+    exportSecretKeyWarning:
+      'Warning: Never share your secret key with anyone. Anyone with your secret key can fully control your account.',
+    showExportSection: 'Export Secret Key (Dangerous)',
+    hideExportSection: 'Hide Export Section',
+    exportWarningFinal:
+      'Final Warning: Sharing this secret key can lead to your account being compromised. Use only for backup purposes.',
+    confirmExport: 'Export Secret Key (nsec format)',
+    yourSecretKey: 'Your Secret Key:',
+    noKeyToExport: 'No key to export. Please check your login status.',
+    localStorage: {
+      title: 'Local Storage',
+      description:
+        'Clear all stored information. You will need to login again after this operation.',
+      clear: 'Clear Information',
+      cleared: 'Local storage cleared',
+    },
+    cacheSettings: {
+      title: 'Secret Caching Settings',
+      description:
+        'Configure whether to temporarily store the secret key in memory. When enabled, you can skip re-authentication within the cache duration, but keeping it for a long time increases the risk of secret key leakage and impersonation.',
+      enabled: 'Cache secret key',
+      disabled: 'Do not cache secret key',
+      saved: 'Settings saved',
+      timeoutLabel: 'Cache Duration (seconds)',
+      clearTitle: 'Clear Cache',
+      clearDescription:
+        'Clear the current secret key cache. You will need to re-authenticate on next operation.',
+      clearButton: 'Clear Cache',
+      clearSuccess: 'Secret key cache has been cleared.',
+      clearError: 'Failed to clear cache.',
+    },
+    logout: {
+      title: 'Logout',
+      description:
+        'Log out of the app. Your keys stay on this device, so you can re-login from the list on the auth screen.',
+      button: 'Logout',
+      success: 'Logout completed.',
+    },
+    appInfo: {
+      title: 'Application Info',
+      version: 'Version:',
+      buildDate: 'Build Date:',
+      commitHash: 'Hash:',
+      repository: 'GitHub:',
+    },
+    language: {
+      title: 'Language Settings',
+      selectLanguage: 'Select language:',
+      japaneseLabel: 'Japanese (日本語)',
+      englishLabel: 'English',
+      changed: 'Language changed',
+    },
+    theme: {
+      title: 'Theme Settings',
+      description: 'Change the appearance of the application.',
+      purpleDark: 'Dark Purple',
+      purpleLight: 'Light Purple',
+      neutralDark: 'Dark Neutral',
+      neutralLight: 'Light Neutral',
+      auto: 'Auto',
+      changed: 'Theme changed',
+    },
+    relays: {
+      title: 'Relay Settings',
+      description:
+        'Relays returned to parent apps via getRelays(). Nosskey itself does not publish to relays.',
+      addPlaceholder: 'wss://relay.example',
+      addButton: 'Add',
+      readLabel: 'Read',
+      writeLabel: 'Write',
+      removeButton: 'Remove',
+      invalidUrl: 'Please enter a URL starting with wss:// or ws://.',
+      empty: 'No relays configured.',
+    },
+    trustedOrigins: {
+      title: 'Trusted sites',
+      description:
+        'Sites and methods you marked as "always allow" in the consent dialog. Trust is recorded per method; other methods are confirmed separately. Removing an entry will make the dialog reappear on the next request.',
+      empty: 'No trusted sites yet.',
+      removeButton: 'Remove',
+      removeAllButton: 'Remove',
+      confirmRemove: 'Remove auto-approval for this method?',
+      confirmRemoveAll: 'Remove auto-approval for this site entirely?',
+    },
+    consentPolicy: {
+      title: 'Per-method consent policy',
+      description:
+        'Set the default behavior for each request type. "Always allow" skips the dialog, "Deny" rejects without prompting.',
+      methodLabel: {
+        connect: 'Site connection (read public key & relay list)',
+        signEvent: 'Sign event (signEvent)',
+        nip44: 'NIP-44 encrypt / decrypt',
+        nip04: 'NIP-04 encrypt / decrypt (legacy)',
+      },
+      option: {
+        ask: 'Ask every time',
+        always: 'Always allow',
+        deny: 'Deny',
+      },
+      denyCount: 'Auto-rejected {count} request(s) this session.',
+      resetDenyCounts: 'Reset deny counter',
+      saved: 'Saved.',
+      corruptionWarning:
+        'Some stored consent settings were invalid and have been reset to defaults. Please verify your settings have not been altered unexpectedly.',
+      decryptTip:
+        'For safety, decryption (NIP-44 / NIP-04 decrypt) always prompts even when set to "Always allow". "Always allow" applies only to non-decrypt operations such as encryption.',
+    },
+    exportKeyInfo: {
+      title: 'Export KeyInfo',
+      description:
+        'You can backup your KeyInfo to use on another device or restore it if your browser data is erased.',
+      warning:
+        'Warning: Losing this KeyInfo may result in loss of access to your account. Store it in a safe place.',
+      showExportSection: 'Export KeyInfo',
+      hideExportSection: 'Hide Export Section',
+      restoreWarning:
+        'This KeyInfo file can be used for recovery when you cannot login with the same passkey and the same username.',
+      exportButton: 'Export KeyInfo',
+      backupData: 'KeyInfo Backup Data:',
+      saveFile: 'Save',
+      saveFileTitle: 'Save to file',
+      noCurrentKeyInfo: 'Current KeyInfo not found. Please check your login status.',
+    },
+    developer: {
+      title: 'For Developers',
+      description: 'PRF extension support check is for debugging purposes.',
+      checkPrf: 'Check PRF Extension Support',
+      prfSupported: 'PRF extension is supported',
+      prfUnsupportedTitle: 'PRF Extension Not Supported',
+    },
+    import: {
+      title: 'Login with KeyInfo',
+      description: 'Use a previously exported KeyInfo file or data to login.',
+      fileSelect: '📁Select KeyInfo File',
+      dataInput: 'Enter KeyInfo Data',
+      dataPlaceholder: 'Paste your Key data here',
+      loginButton: 'Login with Key Data',
+      processing: 'Processing...',
+      or: 'or',
+    },
+  },
+  navigation: {
+    account: 'Account',
+    key: 'Key',
+    settings: 'Settings',
+  },
+  consent: {
+    title: 'Signing request',
+    titleConnect: 'Connection request',
+    titleEncrypt: 'Encryption request',
+    titleDecrypt: 'Decryption request',
+    connectDescription: 'This site will be able to read your public key (npub) and relay settings.',
+    origin: 'Requesting origin:',
+    eventKind: 'Event kind',
+    eventContent: 'Content',
+    eventTags: 'Tags',
+    noTags: '(no tags)',
+    approve: 'Approve',
+    approveOnce: 'Allow once',
+    alwaysAllow: 'Always allow',
+    reject: 'Reject',
+    method: 'Request type',
+    methodLabel: {
+      getPublicKey: 'Read public key',
+      getRelays: 'Read relay list',
+      signEvent: 'Sign event',
+      nip44Encrypt: 'NIP-44 encrypt',
+      nip44Decrypt: 'NIP-44 decrypt',
+      nip04Encrypt: 'NIP-04 encrypt (legacy)',
+      nip04Decrypt: 'NIP-04 decrypt (legacy)',
+    },
+    peerPubkey: 'Peer public key',
+    plaintext: 'Plaintext',
+    decryptNoPreview: 'Ciphertext contents are only visible after decryption.',
+    decryptAlwaysAsk:
+      'For safety, decryption cannot be set to "always allow". This prompt appears on every request.',
+    alwaysAllowSite: 'Always allow this site (skip the consent dialog)',
+    showRaw: 'Show raw event JSON',
+    kindLabel: {
+      metadata: 'Profile (metadata)',
+      textNote: 'Text note',
+      follows: 'Follow list',
+      legacyDm: 'Direct message (NIP-04 legacy)',
+      repost: 'Repost',
+      reaction: 'Reaction',
+      channelMessage: 'Channel message',
+      rumor: 'NIP-17 rumor',
+      seal: 'NIP-17 seal',
+      giftWrap: 'Gift wrap',
+      longForm: 'Long-form article',
+      unknown: 'Other event',
+    },
+    approveAndTrust: 'Approve and always allow',
+    openSettings: 'Open consent settings',
+  },
+  iframeHost: {
+    running: 'Nosskey iframe is running and waiting for signing requests from the parent app.',
+    noKey:
+      'No key configured. Use the button below to open the setup page, then register or sign in with a passkey.',
+    partitionedWarning:
+      'This page is embedded from a different domain. Grant storage access with the button below to reach your key.',
+    grantStorageAccess: 'Grant storage access',
+    storageAccessGranted: 'Access granted. Key loaded.',
+    storageAccessDenied:
+      'Storage access was denied. Retry, or open nosskey.app in another tab to create or unlock your key.',
+    storageAccessUnsupported: 'This browser does not support the Storage Access API.',
+    retry: 'Retry',
+    partitionedTitle: 'Storage access required',
+    deniedTitle: 'Access denied',
+    grantedTitle: 'Access granted',
+    noKeyTitle: 'No key found',
+    unsupportedTitle: 'Browser not supported',
+    openSetup: 'Open setup',
+  },
+};
