+++ examples/svelte-app/src/i18n/translations/ja.ts
@@ -1,0 +1,286 @@
+import type { TranslationData } from './types.js';
+
+// 日本語翻訳
+export const ja: TranslationData = {
+  common: {
+    save: '保存',
+    cancel: 'キャンセル',
+    close: '閉じる',
+    ok: 'OK',
+    error: 'エラー',
+    success: '成功',
+    back: '戻る',
+    copy: 'コピー',
+    copied: 'コピーしました',
+    loading: '読み込み中...',
+    edit: '編集',
+    help: 'ヘルプ',
+    errorMessages: {
+      init: '初期化エラー:',
+      prfCheck: 'PRF対応確認エラー:',
+      passkeyCreation: 'パスキー作成エラー:',
+      login: 'ログインエラー:',
+      importNsec: 'nsec インポートエラー:',
+      prfUnsupported:
+        'このパスキーは利用できません。Bitwarden など一部の認証器は未対応なため、PRF 拡張に対応した別のパスキー（プラットフォームのパスキーなど）でお試しください。',
+    },
+  },
+  appWarning: {
+    title: '注意事項',
+    prfCompatibility: '一部の端末・環境はサポートされていません。',
+  },
+  auth: {
+    title: 'Nosskey',
+    subtitle: 'Nostrアカウントをパスキーで',
+    createNew: '作成',
+    loginWith: 'パスキーでログイン',
+    loading: 'ロード中...',
+    username: 'ユーザー名',
+    usernamePlaceholder: 'ユーザー名を入力',
+    tabLogin: 'ログイン',
+    tabRegister: '新規登録',
+    usernameTip: 'パスキーの表示名として使われます。任意項目で、未入力でも問題ありません。',
+    methodImport: '秘密鍵のインポート（ベータ版）',
+    nsecLabel: '既存の nsec',
+    nsecPlaceholder: 'nsec1...',
+    nsecTip:
+      '既存の Nostr 秘密鍵 (nsec1...) を入力。パスキーで暗号化保存され、署名・暗号化は透過的に動作します。',
+    importNsec: 'インポート',
+    invalidNsec: '無効な nsec 形式です。"nsec1..." で始まる bech32 文字列を入力してください。',
+    accounts: {
+      title: '保存済みアカウント',
+      relogin: 'このアカウントで再ログイン',
+      delete: '削除',
+      confirmDelete: '本当に削除する',
+      cancel: 'キャンセル',
+    },
+  },
+  nostr: {
+    publicKey: '公開鍵',
+    copyToClipboard: 'クリップボードにコピー',
+    copiedToClipboard: 'コピーしました',
+    profileAvatarAlt: 'プロフィール画像',
+    profileFallbackAlt: 'アバター未設定',
+  },
+  key: {
+    title: '鍵管理',
+  },
+  settings: {
+    title: '設定',
+    exportSecretKey: '秘密鍵のエクスポート',
+    exportSecretKeyWarning:
+      '警告：秘密鍵は誰とも共有しないでください。秘密鍵を持つ人はあなたのアカウントを完全に制御できます。',
+    showExportSection: '秘密鍵をエクスポート（危険）',
+    hideExportSection: 'エクスポートセクションを隠す',
+    exportWarningFinal:
+      '最終警告：この秘密鍵を共有すると、あなたのアカウントが乗っ取られる可能性があります。必ずバックアップ目的でのみ使用してください。',
+    confirmExport: '秘密鍵をエクスポートする（nsec形式）',
+    yourSecretKey: 'あなたの秘密鍵：',
+    noKeyToExport: 'エクスポートする鍵がありません。ログイン状態を確認してください。',
+    localStorage: {
+      title: 'ローカルストレージ',
+      description:
+        '保存された情報を全てクリアします。この操作を行うと再度ログインが必要になります。',
+      clear: '情報をクリア',
+      cleared: 'ローカルストレージをクリアしました',
+    },
+    cacheSettings: {
+      title: 'シークレットキャッシュ設定',
+      description:
+        '秘密鍵を一時的にメモリに保存するかどうかを設定します。有効にするとキャッシュ時間内の再認証をスキップできますが、長時間保持すると秘密鍵漏洩やなりすましのリスクが高まります',
+      enabled: '秘密鍵をキャッシュする',
+      disabled: '秘密鍵をキャッシュしない',
+      saved: '設定を保存しました',
+      timeoutLabel: 'キャッシュ時間（秒）',
+      clearTitle: 'キャッシュのクリア',
+      clearDescription:
+        '現在のシークレットキーのキャッシュをクリアします。次回の操作時に再認証が必要になります。',
+      clearButton: 'キャッシュをクリア',
+      clearSuccess: 'シークレットキーのキャッシュがクリアされました。',
+      clearError: 'キャッシュのクリアに失敗しました。',
+    },
+    logout: {
+      title: 'ログアウト',
+      description:
+        'アプリからログアウトします。鍵はこの端末に保持され、認証画面の一覧から再ログインできます。',
+      button: 'ログアウト',
+      success: 'ログアウトが完了しました。',
+    },
+    appInfo: {
+      title: 'アプリケーション情報',
+      version: 'バージョン:',
+      buildDate: 'ビルド日時:',
+      commitHash: 'ハッシュ:',
+      repository: 'GitHub:',
+    },
+    language: {
+      title: '言語設定',
+      selectLanguage: '表示言語を選択:',
+      japaneseLabel: '日本語',
+      englishLabel: '英語 (English)',
+      changed: '言語を変更しました',
+    },
+    theme: {
+      title: 'テーマ設定',
+      description: 'アプリケーションの外観を変更できます。',
+      purpleDark: 'ダークパープル',
+      purpleLight: 'ライトパープル',
+      neutralDark: 'ニュートラルダーク',
+      neutralLight: 'ニュートラルライト',
+      auto: '自動',
+      changed: 'テーマを変更しました',
+    },
+    relays: {
+      title: 'リレー設定',
+      description:
+        '親アプリから getRelays() で参照されるリレー一覧です。Nosskey 自身はリレーへ送信しません。',
+      addPlaceholder: 'wss://relay.example',
+      addButton: '追加',
+      readLabel: '読み取り',
+      writeLabel: '書き込み',
+      removeButton: '削除',
+      invalidUrl: 'wss:// または ws:// で始まる URL を入力してください。',
+      empty: 'リレーが登録されていません。',
+    },
+    trustedOrigins: {
+      title: '信頼済みサイト',
+      description:
+        '同意ダイアログで「このサイトを常に許可」を選んだサイトとメソッドの一覧です。許可はメソッドごとに記録され、他のメソッドは別途確認されます。削除すると次回から再度確認が表示されます。',
+      empty: '信頼済みサイトはまだありません。',
+      removeButton: '削除',
+      removeAllButton: '削除',
+      confirmRemove: 'このメソッドの自動許可を解除しますか？',
+      confirmRemoveAll: 'このサイトの自動許可をすべて解除しますか？',
+    },
+    consentPolicy: {
+      title: 'メソッド別の同意ポリシー',
+      description:
+        'リクエスト種別ごとに既定の挙動を設定できます。「常に許可」はダイアログをスキップ、「拒否」は即拒否します。',
+      methodLabel: {
+        connect: 'サイト接続（公開鍵・リレー設定の読み取り）',
+        signEvent: 'イベント署名 (signEvent)',
+        nip44: 'NIP-44 暗号化 / 復号',
+        nip04: 'NIP-04 暗号化 / 復号（レガシー）',
+      },
+      option: {
+        ask: '毎回確認',
+        always: '常に許可',
+        deny: '拒否',
+      },
+      denyCount: 'このセッションで {count} 件を自動拒否しました。',
+      resetDenyCounts: '拒否カウンタをリセット',
+      saved: '保存しました。',
+      corruptionWarning:
+        '保存された同意設定の一部が不正な形式だったため、デフォルト値に戻しました。意図しない設定変更がないか確認してください。',
+      decryptTip:
+        '安全のため、復号（NIP-44 / NIP-04 decrypt）は「常に許可」にしても毎回確認します。「常に許可」は暗号化など復号以外にのみ適用されます。',
+    },
+    exportKeyInfo: {
+      title: '鍵情報のエクスポート',
+      description:
+        '鍵情報をバックアップして別のデバイスで利用したり、ブラウザデータが消去された場合に復元することができます。',
+      warning:
+        '注意: この鍵情報を紛失するとアカウントへのアクセスができなくなる場合があります。安全な場所に保管してください。',
+      showExportSection: '鍵情報をエクスポート',
+      hideExportSection: 'エクスポートセクションを隠す',
+      restoreWarning:
+        'この鍵情報ファイルは、同じパスキーと同じユーザー名でログインできない場合の復元に使用できます。',
+      exportButton: '鍵情報をエクスポート',
+      backupData: '鍵情報のバックアップデータ:',
+      saveFile: '保存',
+      saveFileTitle: 'ファイルに保存',
+      noCurrentKeyInfo: '現在の鍵情報が見つかりません。ログイン状態を確認してください。',
+    },
+    developer: {
+      title: '開発者向け',
+      description: 'PRF拡張対応の確認はデバッグ用途です。',
+      checkPrf: 'PRF拡張対応確認',
+      prfSupported: 'PRF拡張がサポートされています',
+      prfUnsupportedTitle: 'PRF拡張がサポートされていません',
+    },
+    import: {
+      title: '鍵情報でログイン',
+      description: '以前にエクスポートした鍵情報ファイルまたはデータを使用してログインします。',
+      fileSelect: '📁鍵情報ファイルを選択',
+      dataInput: '鍵情報データを入力',
+      dataPlaceholder: '鍵情報データをここに貼り付けてください',
+      loginButton: '鍵情報データでログイン',
+      processing: '処理中...',
+      or: 'または',
+    },
+  },
+  navigation: {
+    account: 'アカウント',
+    key: '鍵管理',
+    settings: '設定',
+  },
+  consent: {
+    title: '署名リクエストの確認',
+    titleConnect: '接続リクエストの確認',
+    titleEncrypt: '暗号化リクエストの確認',
+    titleDecrypt: '復号リクエストの確認',
+    connectDescription:
+      'このサイトはあなたの公開鍵（npub）とリレー設定を読み取れるようになります。',
+    origin: 'リクエスト元:',
+    eventKind: 'イベント種別',
+    eventContent: '本文',
+    eventTags: 'タグ',
+    noTags: '(タグなし)',
+    approve: '承認',
+    approveOnce: '今回のみ許可',
+    alwaysAllow: '常に許可',
+    reject: '拒否',
+    method: 'リクエスト種別',
+    methodLabel: {
+      getPublicKey: '公開鍵の取得',
+      getRelays: 'リレー設定の取得',
+      signEvent: 'イベント署名',
+      nip44Encrypt: 'NIP-44 暗号化',
+      nip44Decrypt: 'NIP-44 復号',
+      nip04Encrypt: 'NIP-04 暗号化（レガシー）',
+      nip04Decrypt: 'NIP-04 復号（レガシー）',
+    },
+    peerPubkey: '相手の公開鍵',
+    plaintext: '平文',
+    decryptNoPreview: '暗号文の内容は復号後にしか確認できません。',
+    decryptAlwaysAsk:
+      '安全のため、復号は「常に許可」にできません。リクエストのたびにこの確認が表示されます。',
+    alwaysAllowSite: 'このサイトを常に許可（同意ダイアログをスキップ）',
+    showRaw: '生のイベント JSON を表示',
+    kindLabel: {
+      metadata: 'プロフィール (metadata)',
+      textNote: 'テキストノート',
+      follows: 'フォローリスト',
+      legacyDm: 'ダイレクトメッセージ（NIP-04 レガシー）',
+      repost: 'リポスト',
+      reaction: 'リアクション',
+      channelMessage: 'チャンネルメッセージ',
+      rumor: 'NIP-17 Rumor',
+      seal: 'NIP-17 Seal',
+      giftWrap: 'Gift Wrap',
+      longForm: '長文記事',
+      unknown: 'その他のイベント',
+    },
+    approveAndTrust: '常に許可して承認',
+    openSettings: '同意設定を開く',
+  },
+  iframeHost: {
+    running: 'Nosskey iframe が起動中です。親アプリからの署名リクエストを待機します。',
+    noKey:
+      '鍵が設定されていません。下のボタンからセットアップ画面を開き、パスキーで登録またはログインしてください。',
+    partitionedWarning:
+      'このサイトは別ドメインから埋め込まれています。鍵にアクセスするには、下のボタンを押してストレージアクセスを許可してください。',
+    grantStorageAccess: 'ストレージアクセスを許可',
+    storageAccessGranted: 'アクセスが許可されました。鍵を読み込みました。',
+    storageAccessDenied:
+      'ストレージアクセスが拒否されました。もう一度試すか、別タブで nosskey.app を開いてください。',
+    storageAccessUnsupported: 'このブラウザは Storage Access API をサポートしていません。',
+    retry: '再試行',
+    partitionedTitle: 'ストレージアクセスの許可が必要です',
+    deniedTitle: 'アクセスが拒否されました',
+    grantedTitle: '許可されました',
+    noKeyTitle: '鍵が見つかりません',
+    unsupportedTitle: 'このブラウザは未対応です',
+    openSetup: 'セットアップを開く',
+  },
+};
